% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModel.R
\name{dash.assign.MetaModel}
\alias{dash.assign.MetaModel}
\title{\verb{$<-}}
\usage{
dash.assign.MetaModel(x, name, value)
}
\arguments{
\item{x}{A MetaModel subclass' instance.}

\item{name}{The name of the attribute/method}

\item{value}{The name of the attribute/method}
}
\value{
The model's attributes or method.
}
\description{
Assign the value to the field of the model instance.
If the "lazy" parameter of the meta attribute is TRUE and the attribute
is a ModelField, the value is queried from the database and stored in
the object.
If the value is already stored in the ModelField, the attribute is
directly returned.
}
