% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaModel.R
\name{dash.MetaModel}
\alias{dash.MetaModel}
\title{\code{$}}
\usage{
dash.MetaModel(x, name)
}
\arguments{
\item{x}{A MetaModel subclass' instance.}

\item{name}{The name of the attribute/method}
}
\value{
The model's attributes or method.
}
\description{
Returns asked method or attribute's value.
If the "lazy" parameter of the meta attribute is TRUE and the attribute
is a ModelField, the value is queried from the database and stored in
the object.
If the value is already stored in the ModelField, the attribute is
directly returned.
}
