% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ModelBuilder}
\alias{ModelBuilder}
\title{ModelBuilder}
\usage{
ModelBuilder(.orm__)
}
\arguments{
\item{.orm__}{the curent orm to attach the model to.}
}
\description{
This method is a helper to create models.
The first parameter is the name of the table.
Other parameters must be named and hold a ModelField instance, like
this:

Person <- ORM$ModelBuilder(
"Person",
name=CharacterField(),
nickname=CharacterField(nullable=TRUE),
adress=ForeignKeyField(Adress, type=MANY_TO_MANY, nullable=TRUE)
)

The model can then be used to manipulate the database's content.
See the MetaModel help for more informations:
?MetaModel
}
